Add --force option to ipactl
If an error occurs in the start up sequence in ipactl start/restart,
all the services are stopped. Using the --force option prevents
stopping of services that have successfully started, just skips the
services which can not be started.

ipactl status now shows stopped services also, if the directory
server is running.

With the contribution of Ana Krivokapic

https://fedorahosted.org/freeipa/ticket/3509

Reviewed-By: Martin Kosek <mkosek@redhat.com>